All user related API calls now pass sunshine tests
All the user related API calls now pass their tests, with the exception
of the get user range, as I am waiting for some answers there.

+++ src/main/java/no/ntnu/imt3281/ludo/common/MessageUtility.java
@@ -10,7 +10,7 @@ public class MessageUtility {
         var error = new JSONObject();
         error.put(FieldNames.ID, requestID);
         var codes = new JSONArray();
-        codes.put(code);
+        codes.put(code.toInt());
         error.put(FieldNames.CODE, codes);
         errors.put(error);
     }
